Add discussion and comment system to ReadBook page
Introduces a full-featured discussion and comment system for books, including threaded comments, emoji support, reporting, and admin moderation. Adds new components for discussion lists, views, forms, modals, and menus. Integrates @rails/actioncable for real-time updates, updates dependencies, and improves navigation logic for book access.
