+++ src/Pages/ReadBook/components/Discussion/Icons.tsx
@@ -1,0 +1,14 @@
+
+export const ChevronLeftIcon = ({ className }: { className?: string }) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    viewBox="0 0 24 24"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="2"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+    className={className}
+  >
+  </svg>
+);
